test: extract PageTitle component and test
