+++ src/features/list/UserList.module.css
@@ -4,10 +4,6 @@
   align-items: center;
 }
 
-.pageTitle {
-  composes: pageTitle from "../../shared/style/shared.module.css";
-}
-
 .paginatedBlock {
   display: flex;
 }
