+++ src/shared/component/PageTitle.module.css
@@ -1,0 +1,3 @@
+.pageTitle {
+  text-align: center;
+}
