+++ src/shared/component/index.js
@@ -1,0 +1,5 @@
+import Navigator from "./Navigator";
+import PageTitle from "./PageTitle";
+import Pending from "./Pending";
+
+export { Navigator, PageTitle, Pending };
