--- src/shared/style/shared.module.css
@@ -1,3 +1,0 @@
-.pageTitle {
-  text-align: center;
-}
